separate into components, cleanup, adjust app color
