+++ src/App.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import DogSearch from './components/DogSearch.vue'
-import { onMounted, ref } from 'vue'
+import { onMounted } from 'vue'
 import { useDogStore } from './stores/dog'
 import { useAuthStore } from './stores/auth'
 
@@ -8,14 +8,6 @@ const dogStore = useDogStore()
 const authStore = useAuthStore()
 onMounted(async () => {
   await dogStore.getBreeds()
-})
-
-const hasFavorites = ref(false)
-dogStore.$subscribe((_, store) => {
-  const favorites = Object.keys(store.favoritesList)
-  if (favorites.length > 0) {
-    hasFavorites.value = true
-  } else hasFavorites.value = false
 })
 </script>
 
@@ -39,7 +31,7 @@ dogStore.$subscribe((_, store) => {
           type="text"
           name="name"
           id="name"
-          class="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
+          class="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-cyan-600 sm:text-sm sm:leading-6"
           placeholder="Name"
           v-model="authStore.name"
         />
@@ -52,7 +44,7 @@ dogStore.$subscribe((_, store) => {
           type="email"
           name="email"
           id="email"
-          class="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
+          class="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-cyan-600 sm:text-sm sm:leading-6"
           placeholder="Email"
           v-model="authStore.email"
         />
@@ -60,7 +52,7 @@ dogStore.$subscribe((_, store) => {
 
       <button
         type="button"
-        class="mt-4 rounded-md bg-indigo-500 px-3 py-2 text-sm font-semibold text-white shadow-sm hover:bg-indigo-400 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-500"
+        class="mt-4 rounded-md bg-cyan-600 px-3 py-2 text-sm font-semibold text-white shadow-sm hover:bg-cyan-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-cyan-500"
         @click.prevent="authStore.login"
       >
         Login
@@ -70,18 +62,10 @@ dogStore.$subscribe((_, store) => {
     <div v-else>
       <button
         type="button"
-        class="rounded-md bg-indigo-500 px-3 py-2 text-sm font-semibold text-white shadow-sm hover:bg-indigo-400 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-500"
+        class="rounded-md bg-cyan-600 px-3 py-2 text-sm font-semibold text-white shadow-sm hover:bg-cyan-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-cyan-500"
         @click.prevent="authStore.logout"
       >
         Log Out
-      </button>
-      <button
-        v-if="hasFavorites"
-        type="button"
-        class="ml-3 rounded-md bg-indigo-500 px-3 py-2 text-sm font-semibold text-white shadow-sm hover:bg-indigo-400 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-500"
-        @click.prevent="dogStore.findMatch"
-      >
-        Find Match
       </button>
       <DogSearch />
     </div>
