+++ src/components/DogModal.vue
@@ -87,7 +87,7 @@ function handleClose() {
                         v-model="currentDog.favorite"
                         :class="[
                           currentDog.favorite ? 'bg-green-600' : 'bg-gray-200',
-                          'relative inline-flex h-6 w-11 flex-shrink-0 cursor-pointer rounded-full border-2 border-transparent transition-colors duration-200 ease-in-out focus:outline-none focus:ring-2 focus:ring-indigo-600 focus:ring-offset-2'
+                          'relative inline-flex h-6 w-11 flex-shrink-0 cursor-pointer rounded-full border-2 border-transparent transition-colors duration-200 ease-in-out focus:outline-none focus:ring-2 focus:ring-cyan-600 focus:ring-offset-2'
                         ]"
                         @click="dogStore.toggleFavorite"
                       >
